requirements.txt
flask 
flask-cors
pytesseract 
pillow
numpy<2.0
opencv-python-headless
requests
python-dotenv
